Fix autocorr flow with antenna numbers >= 192
This fixes a bug in `hera_catcher_disk_thread_bda.c` that prevented the auto
correlations from being properly propagated downstream when observing with
antennas whose antenna numbers are greater than or equal to 192.  It also
removes use of the `Nants` and `corr_to_hera_map` variables before they get
initialized.  Read on for more details.

Each block of the output databuf for `hera_catcher_disk_thread_bda` contains a
header that has nants, jd, and an array of bytes used as per-antenna flags to
indicate that a particular antenna has data.  `hera_catcher_disk_thread_bda`
maintains an `auto_ants_filled` counter variable.  When it is 0, a free output
block is acquired and these flags are initialized to 0.  As the auto
correlations are copied into the output buffer, the `auto_ants_filled` counter
is incremented.  When `auto_ants_filled >= Nants` the block is marked as filled
(i.e. handed over to the downstream thread) and `auto_ants_filled` is set back
to 0, setting up for the acquisition of another free output block.

Another detail is that the autos are copied and `auto_ants_filled` is
incremented if and only if the flag for that antenna is 0 (i.e. only if it has
not yet been copied).

The problem is that the flags are indexed by antenna number (NOT correlator
input pair number) but the flags array is sized based on the number of
correlator input pairs.  It appears that the current correlator is built for
192 input pairs (i.e. 0-191), but there are a number of antennas that have
antenna numbers greater than 191.  The header flag for these antennas actually
exists beyond the end of the databuf block's header, in the data portion of of
the block.  The first time through the data buffer, these locations are 0 so
these autos get copied and `auto_ants_filled` gets incremented.  The next time
through the data buffer, these locations likely contain non-zero data values,
so these autos are never copied nor is `auto_ants_filled` incremented for them.
This leads to `auto_ants_filled` never reaching `Nants` so the block is never
marked as filled so the downstream thread is starved for data.