Use AUTOCORR_N_BLOCKS consistently and set to 4
`AUTOCORR_N_BLOCKS` and `CATCHER_N_BLOCKS` are distinctly different (i.e they
serve different purposes) even though their current values are identical.
Prior to this, `hera_catcher_autocorr_databuf_create` used ``CATCHER_N_BLOCKS``
(which is defined to be 4) for the number of blocks to create, but the threads
using this databuf used `AUTOCORR_N_BLOCKS` (which was defined to be 2).  Now
`AUTOCORR_N_BLOCKS` is defined to be 4 and the create functionx
usesa`AUTOCORR_N_BLOCKS`  for the number of blocks to create.