Match type of process_packet's returned bcnt value
The process_packet() function in hera_catcher_ibvpkt_thread.c returns a
bcnt value as a `uint32_t`.  Returning `0xffffffff` and storing it in a
uint64_t variable results in the value `0x00000000ffffffff`, which does
not compare equal to `-1`!  This resulted in locking/updating/unlocking
the status buffer on every packet resulting in poor throughput.