Increase number of working blocks from 2 to 4
With increased data rates, two working blocks proved to be insufficient
to cover the range of incoming bcnts from all X engines.  Ideally the
network fabric would deliver packets in round robin synchronized format
from the X engines such that the incoming bcnt values are very cleanly
structured.  Unfortunamtely the reality is that the incoming bcnts can
have a fair bit of spread to them.  This spread is sometimes referred to
in comments as "bcnt dispersion".

The number of working blocks is now specifed via N_WORKING_BLOCKS rather
than hardcoding numerical constants 3 and 2 (and implicitly 1).