Fix pointer arthmetic bug in hera_ibvpkt_thread.c
In C, pointer arithmetic is in units of the type size (i.e. not bytes,
unless the type size is 1 byte).  In this case, we were adding a byte
count to a uint64_t pointer (`payload_p`) without dividing by
`sizeof(uint64_t)`.  Maybe `sizeof(*payload_p)` would be even better as
it would be immune to changes in the type of `payload p`?