Calculate strt_bcnt/stop_bcnt from header.bcnt[0]
Indexing into header.bcnt using bctr to get strt_bcnt and stop_bcnt is
susceptible is old/stale values in the event of missed packets.
Calculating start_bcnt and stop_bcnt from header.bcnt[0] is more robust
because header.bcnt[0] is populated on the first pacet received for the
block (even if the packet's bcnt is for later in the block).